refactor(chat): extract reusable ModelPicker from ModelSelector
Splits the model dropdown into a presentational 'ModelPicker' ('renderer/src/features/chat/components/model-picker.tsx') and keeps 'ModelSelector' as a thin wrapper that owns the chat-settings-specific glue (provider credential lookup, onOpenSettings wiring). The picker now accepts a plain 'value / onChange / onClear / onOpenSettings' interface so it can be reused outside the chat (agent form, future screens) without duplicating the grouped-by-provider layout and search filter.

Behavior unchanged: same grouping, same filter, same empty state, same provider credential gating.
